+++ core/common/src/commonMain/kotlin/com/chimali/core/common/BuildVariant.kt
@@ -1,0 +1,6 @@
+package com.chimali.core.common
+
+/**
+ * Indicates if the current build is a debug build.
+ */
+expect val isDebug: Boolean