+++ core/common/src/iosMain/kotlin/com/chimali/core/common/BuildVariant.kt
@@ -1,0 +1,8 @@
+package com.chimali.core.common
+
+import kotlin.native.Platform
+
+/**
+ * iOS implementation of build variant detection.
+ */
+actual val isDebug: Boolean = Platform.isDebugBinary